Add configuration options to sample project
- Flags can be used for configuring Connect() in cli args
- ENV vars can be used for configuring Connect()
- moved `osbindings.go` into root directory for ease of importing `onesphere` package
- Add step by step instructions to README as a simple Getting Started guide

resolves #1
